[#2] Refactor of behaviour on power up.
The power up now shoots the ball if when activated the latest is inside
the 'power up area' of the paddle. The shoot power is defined to the
distance between the ball and the center of the paddle. Also, the paddle
can shoot even if the ball already surpassed it.