[#20] Refactor for ball and paddle. Add powerup bar.
Paddle and Ball domains where very bounded and Ball was heavily modifying
the state of Paddle and viceversa in a not safe or cohesive way.

Using signals we can subscribe each object to the fired event and change
its behavior without the need to check internal components or flags from
others objects.

Also, we added a powerup meter for Paddle that depletes when loading
the powerup that also replace the Paddle timer.